Refaktorerer validering av ettersending
